# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Merchant Payment Intelligence - Ingestion Foundation
  version: 0.1.0
  description: 'Merchant Payment Intelligence Platform -- backend API for the frontend
    team.


    Covers Steps 1-5 (ingestion, identity resolution, feature store), the complete
    Fraud/Anomaly Detection engine (Step 6a): Track A (behavioral snapshots), Track
    B (Isolation Forest), Track C (time-series drift, including Funnel Account detection),
    Track D (HDBSCAN clustering), and the Section 8 final aggregation that combines
    all three model tracks into one final_anomaly_score/anomaly_band per entity --
    plus the complete Operational Issues engine (Step 6b, all four detectors: Duplicate
    Payment, Formatting Rejection, Batch Never Settles, Network/Processor Timeout)
    and the Reconciliation engine (Step 6c).


    All `/anomaly/*/compute` and `/resolve/parties` and `/features/compute` endpoints
    are POST, take an optional `{"tenant_bank_id": null}` body (omit or null = all
    tenants), and return a run-summary object, not the computed rows themselves --
    fetch those via the matching GET endpoint. Run them in dependency order for a
    fresh tenant: snapshots/compute -> isolation-forest/train + timeseries/compute
    + clustering/compute (any order) -> final-score/compute.'
  contact:
    name: Vishaal
    email: vishaal.g@tibilsolutions.com
paths:
  /ingest/file:
    post:
      summary: Ingest File
      operationId: ingest_file_ingest_file_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_ingest_file_ingest_file_post'
        required: true
      responses:
        '200':
          description: File ingested (even if some rows failed -- check rows_failed/errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestFileResult'
        '422':
          description: Missing/invalid form fields.
        '400':
          description: Wrong file extension (must be .csv, .xlsx, or .xls).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
      - Ingestion (Steps 1-3)
  /resolve/parties:
    post:
      summary: Resolve Parties Endpoint
      operationId: resolve_parties_endpoint_resolve_parties_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolvePartiesRequest'
              default: {}
      responses:
        '200':
          description: Resolution run completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolvePartiesResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Identity Resolution (Step 4)
  /merchants:
    get:
      summary: List Merchants
      operationId: list_merchants_merchants_get
      parameters:
      - name: tenant_bank_id
        in: query
        required: true
        schema:
          type: string
          title: Tenant Bank Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Page of merchants for this tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Identity Resolution (Step 4)
  /merchants/{merchant_id}:
    get:
      summary: Get Merchant
      operationId: get_merchant_merchants__merchant_id__get
      parameters:
      - name: merchant_id
        in: path
        required: true
        schema:
          type: string
          title: Merchant Id
      responses:
        '200':
          description: Merchant found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: No merchant with this merchant_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
      - Identity Resolution (Step 4)
  /individuals:
    get:
      summary: List Individuals
      operationId: list_individuals_individuals_get
      parameters:
      - name: tenant_bank_id
        in: query
        required: true
        schema:
          type: string
          title: Tenant Bank Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Page of individuals for this tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Identity Resolution (Step 4)
  /individuals/{individual_id}:
    get:
      summary: Get Individual
      operationId: get_individual_individuals__individual_id__get
      parameters:
      - name: individual_id
        in: path
        required: true
        schema:
          type: string
          title: Individual Id
      responses:
        '200':
          description: Individual found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: No individual with this individual_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
      - Identity Resolution (Step 4)
  /features/compute:
    post:
      summary: Compute Features Endpoint
      operationId: compute_features_endpoint_features_compute_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComputeFeaturesRequest'
              default: {}
      responses:
        '200':
          description: Feature computation run completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputeFeaturesResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Feature Store (Step 5, dashboard only)
  /features:
    get:
      summary: List Features
      operationId: list_features_features_get
      parameters:
      - name: tenant_bank_id
        in: query
        required: true
        schema:
          type: string
          title: Tenant Bank Id
      - name: party_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Party Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Page of party feature rollups.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyFeaturesList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Feature Store (Step 5, dashboard only)
  /features/{party_id}:
    get:
      summary: Get Features
      operationId: get_features_features__party_id__get
      parameters:
      - name: party_id
        in: path
        required: true
        schema:
          type: string
          title: Party Id
      responses:
        '200':
          description: Features found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyFeatures'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          description: No features computed yet for this party_id -- run POST /features/compute
            first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      tags:
      - Feature Store (Step 5, dashboard only)
  /anomaly/snapshots/compute:
    post:
      summary: Compute Snapshots Endpoint
      operationId: compute_snapshots_endpoint_anomaly_snapshots_compute_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComputeSnapshotsRequest'
              default: {}
      responses:
        '200':
          description: Snapshot computation run completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputeSnapshotsResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Anomaly Detection - Behavioral Snapshots (Track A)
  /anomaly/isolation-forest/train:
    post:
      summary: Train Isolation Forest Endpoint
      operationId: train_isolation_forest_endpoint_anomaly_isolation_forest_train_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainIsolationForestRequest'
              default: {}
      responses:
        '200':
          description: Training run completed for both segments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsolationForestTrainResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Anomaly Detection - Isolation Forest (Track B)
...
components:
  schemas:
    Body_ingest_file_ingest_file_post:
      properties:
        tenant_bank_id:
          type: string
          title: Tenant Bank Id
        rail_type:
          type: string
          title: Rail Type
        settlement_stage:
          $ref: '#/components/schemas/SettlementStage'
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - tenant_bank_id
      - rail_type
      - settlement_stage
      - file
      title: Body_ingest_file_ingest_file_post
...
    HTTPError:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
...
servers:
- url: http://localhost:8000
  description: Local dev (uvicorn app.main:app --reload)
...